char: sbt_lockamp: Fix runtime PM-related issue
On the Green Mango platform, the PMIC's VLDO10 is turned off when the
lock-in amplifier powers down (even at runtime). In turn, this causes the
ADC and DAC to go offline. For some reason, this corrupts the internal
state in the FPGA so that, e.g., the MA filter doesn't work as intended.

We fix the abovementioned issue by always resetting the Lock-in amplifier
(via the GPIO reset pin) whenever the driver wakes up.

Indirectly, this may also fix other data corruption issues. Time will
tell.