char: sbt_lockamp: Restore HW registers after power off
This is accomplished with the caching feature of regmap.

This fixes an issue with the recent reset added after power off. Specifically,
after the reset, all registers had default values. Now, the registers are
restored to the values they had before the power off.

This is a large commit due to the addition of the regmap. There is also assorted
cleanup/refactoring.