char: sbt_lockamp: Fix power management issues with 'amp_supply_force_off'
Specifically, the amp_supply regulator was sometimes erroneously enabled two times
but only disabled once. This caused the regulator to stay enabled even
though it was supposed to turn off.

The problem is the complex relationship between the runtime power
management and the 'amp_supply_force_off' feature.

Now, the amp_supply regulator is owned exclusively by this driver. In turn,
this enables us to use the regulator_is_enabled function to query the
state of the regulator and enable/disable it accordingly. This greatly
reduces the margin of error in the combination of 'amp_supply_force_off'
and runtime power management.