spi: spi-xilinx: fix race condition in interrupt handler
We now set the XSPI_CR_TRANS_INHIBIT bit for the duration of the
interrupt handler. This way, we ensure that the FPGA does not
interfere with the CPU's read/write of the Rx/Tx FIFOs.

In practice, this fixes kernel panics such as:

 * Unhandled fault: imprecise external abort (0x1406)
 * Internal error: Oops - BUG: 1406 [#1] PREEMPT SMP ARM

Both of the above were sometimes followed by a register and stack
dump belonging to an unrelated process (whatever process the kernel
preempted just before the interrupt handler ran).

Note that the documentation for the "AXI Quad SPI" IP specifically
recommends to set the master inhibit bit:

 > Set the master transaction inhibit bit to service the interrupt
 > request.

See item 10 on page 63 of
https://docs.xilinx.com/v/u/en-US/ds843_axi_quad_spi

Moreover, this was actually the previous behaviour (that is, to set
the inhibit bit during the interrupt) before commit
41fefabaa09fece98786cec4ff22ad052adc1b87 changed things around.