gpio: gpio-zynq: fix power management on IRQ-only use case
A recent commit simplified the irq_request_resources/irq_release_resources functions
away. Doing so also broke the IRQ-only use case. That is, when you only request
IRQs and no GPIOs. After the simplification, only the request/free GPIO functions would
turn on/off power to the chip. This left the power turned off if you did not request
any GPIOs (e.g., if you only request IRQs).

Now (with this commit), we also turn on/off the power on the
irq_request_resources/irq_release_resources functions. This is how things used to be
before the simplification.