mfd: da9063-core: add power off and restart handlers
This enables system with a DA906X chip to do a low-level power off/restart.
That is, to power cycle the entire system through the PMIC as opposed to
just restarting the chip with the CPU.

Note that da9063_wdt already registers a 'restart' handler. However, said
handler actually performs a power off. I intentionally left this handler
unchanged (to not break existing use cases whatever they may be). To work
around the issue, I simply register the new restart handler with higher
priority so that the system chooses it.